requirements.txt
pyyaml
flake8
pylint >= 2.17.0
pytest
cerberus